+++ src/api/store/mercadopago/payment-methods/query-config.ts
@@ -1,0 +1,5 @@
+export const listPaymentmethodsQueryConfig = {
+    defaults: ['account_holder.*'],
+    isList: true,
+    defaultLimit: 50
+}