+++ src/api/store/mercadopago/payment/validators.ts
@@ -7,8 +7,19 @@ export const PostStoreMercadopagoPayment = z.object({
         token: z.string().min(1),
         transaction_amount: z.number().min(1),
         installments: z.number(),
-        payer: z.object({
-            email: z.string(),
-        })
+        payer: z.union([
+            z.object({
+                email: z.string().optional(),
+                identification: z.object({
+                    type: z.string(),
+                    number: z.string()
+                }).optional(),
+            }).strict(),
+            z.object({
+                type: z.string(),
+                id: z.string(),
+            }).strict()
+        ]),
+        payment_method_id: z.string(),
     })
 })