+++ src/providers/mercado-pago/service.ts
@@ -13,27 +13,39 @@ import {
   CancelPaymentOutput,
   CapturePaymentInput,
   CapturePaymentOutput,
+  CreateAccountHolderInput,
+  CreateAccountHolderOutput,
   DeletePaymentInput,
   DeletePaymentOutput,
   GetPaymentStatusInput,
   GetPaymentStatusOutput,
   InitiatePaymentInput,
   InitiatePaymentOutput,
+  ListPaymentMethodsInput,
+  ListPaymentMethodsOutput,
+  PaymentCustomerDTO,
   ProviderWebhookPayload,
   RefundPaymentInput,
   RefundPaymentOutput,
   RetrievePaymentInput,
   RetrievePaymentOutput,
+  SavePaymentMethodInput,
+  SavePaymentMethodOutput,
+  UpdateAccountHolderInput,
+  UpdateAccountHolderOutput,
   UpdatePaymentInput,
   UpdatePaymentOutput,
   WebhookActionResult,
 } from "@medusajs/types";
-import MercadoPagoConfig, { Payment, PaymentRefund } from "mercadopago";
+import MercadoPagoConfig, { Customer, CustomerCard, Payment, PaymentRefund } from "mercadopago";
 import { PaymentCreateRequest } from "mercadopago/dist/clients/payment/create/types";
 import { PaymentSearchResult } from "mercadopago/dist/clients/payment/search/types";
 import { MercadopagoOptions, MercadopagoWebhookPayload } from "../../types";
 import { createHmac } from "crypto";
 import { Logger } from "@medusajs/medusa";
+import { CustomerRequestBody } from "mercadopago/dist/clients/customer/commonTypes";
+import { CustomerUpdateData } from "mercadopago/dist/clients/customer/update/types";
+import { PostStoreMercadopagoPaymentType } from "../../api/store/mercadopago/payment/validators";
 
 type InjectedDependencies = {
   logger: Logger;
@@ -113,7 +125,7 @@ class MercadopagoProviderService extends AbstractPaymentProvider<MercadopagoOpti
     data = results[0] ?? data;
 
     // Returning PaymentSessionStatus.CAPTURED for auto capture, since for UY card method they are captured automatically
-    // Should make it conditionally in cases where a method could be not auto captur ed (maybe cash or another country)
+    // TODO: Should make it conditionally in cases where a method could be not auto captured (maybe cash or another country)
     return {
       data: data,
       status: PaymentSessionStatus.CAPTURED,
@@ -123,12 +135,7 @@ class MercadopagoProviderService extends AbstractPaymentProvider<MercadopagoOpti
   async capturePayment(
     input: CapturePaymentInput
   ): Promise<CapturePaymentOutput> {
-    if (!input.data) {
-      throw new MedusaError(
-        MedusaErrorTypes.INVALID_DATA,
-        "No data found in input for capture payment"
-      );
-    }
+    // For now, we assume auto capture, so no need to do anything else with Mercado Pago
     return { data: input.data };
   }
 
@@ -193,7 +200,8 @@ class MercadopagoProviderService extends AbstractPaymentProvider<MercadopagoOpti
   }
 
   updatePayment(input: UpdatePaymentInput): Promise<UpdatePaymentOutput> {
-    // Until working with preferences, we won't have a need for this method
+    // TODO: Until working with preferences, we won't have a need for this method. After including preferences, it should update preference items and total amount
+    // when items are added / removed from cart
     return Promise.resolve({ data: input.data });
   }
 
@@ -240,12 +248,152 @@ class MercadopagoProviderService extends AbstractPaymentProvider<MercadopagoOpti
     }
   }
 
+  async createAccountHolder?({
+    context
+  }: CreateAccountHolderInput
+  ): Promise<CreateAccountHolderOutput> {
+    const { account_holder, customer, idempotency_key } = context
+
+    const id = account_holder?.data.id as string | undefined
+
+    if (id) {
+      return { id }
+    }
+
+    if (!customer) {
+      throw new MedusaError(MedusaErrorTypes.INVALID_DATA, "No customer provided while creating account holder")
+    }
+
+    try {
+      const customerClient = new Customer(this.client_);
+      const body: CustomerRequestBody = {
+        email: customer.email,
+        first_name: customer.first_name || undefined,
+        last_name: customer.last_name || undefined,
+        date_registered: new Date().toISOString(),
+      };
+
+      const createdCustomer = await customerClient.create({
+        body: body,
+        requestOptions: {
+          idempotencyKey: idempotency_key,
+        },
+      });
+
+      return {
+        id: createdCustomer.id!,
+        data: createdCustomer as unknown as Record<string, unknown>
+      }
+    } catch (error) {
+      throw new MedusaError(MedusaErrorTypes.UNEXPECTED_STATE, "An error occurred while trying to create a Mercado Pago customer")
+    }
+  }
+
+  async updateAccountHolder({
+    context,
+  }: UpdateAccountHolderInput): Promise<UpdateAccountHolderOutput> {
+    const { account_holder, customer, idempotency_key } = context
+
+    const accountHolderId = account_holder.data?.id as string | undefined
+
+    if (!accountHolderId) {
+      throw new MedusaError(MedusaErrorTypes.INVALID_DATA, "No account holder provided while updating account holder")
+    }
+
+    // No customer, nothing to upodate with third party
+    if (!customer) {
+      return {}
+    }
+
+    try {
+      const customerClient = new Customer(this.client_);
+      const body: CustomerRequestBody = {
+        first_name: customer.first_name ?? undefined,
+        last_name: customer.last_name ?? undefined,
+        phone: { number: customer.phone ?? undefined },
+        identification: account_holder.data.identification ?? undefined,
+      }
+      if (!account_holder.data.email) {
+        body.email = customer.email
+      }
+
+      const payload: CustomerUpdateData = {
+        customerId: accountHolderId,
+        body: body,
+        requestOptions: {
+          idempotencyKey: idempotency_key,
+        }
+      } 
+
+      const updatedCustomer = await customerClient.update(payload);
+
+      return {
+        data: updatedCustomer as unknown as Record<string, unknown>,
+      }
+    } catch (e) {
+      throw new MedusaError(MedusaErrorTypes.UNEXPECTED_STATE, "An error occurred in updateAccountHolder when updating a Stripe customer")
+    }
+  }
+
+  async savePaymentMethod({
+    context,
+    data,
+  }: SavePaymentMethodInput): Promise<SavePaymentMethodOutput> {
+    const accountHolderId = context?.account_holder?.data?.id as
+      | string
+      | undefined
+
+    if (!accountHolderId) {
+      throw new MedusaError(MedusaErrorTypes.INVALID_DATA, "Account holder not set while saving a payment method")
+    }
+
+    const paymentMethodData = data as PaymentCreateRequest
+
+    const card = new CustomerCard(this.client_)
+
+    // I opened a support ticket with Mercado Pago, as this call randomly fails or succeeds
+    // for now, the step that calls this method has a try / catch, to allow the Payment to cotinue even if this fails
+    const created = await card.create({
+      customerId: accountHolderId,
+      body: { 
+        token: paymentMethodData.token, 
+      },
+      requestOptions: {
+        idempotencyKey: context?.idempotency_key
+      }
+    })
+
+    return { id: created.id!, data: created as unknown as Record<string, unknown> }
+  }
+
+  async listPaymentMethods({
+    context,
+  }: ListPaymentMethodsInput): Promise<ListPaymentMethodsOutput> {
+    const accountHolderId = context?.account_holder?.data?.id as
+      | string
+      | undefined
+    if (!accountHolderId) {
+      return []
+    }
+
+    const cardClient = new CustomerCard(this.client_)
+
+    const paymentMethods = await cardClient.list({
+      customerId: accountHolderId
+    })
+
+    return paymentMethods.map((method) => ({
+      id: method.id!,
+      data: method as unknown as Record<string, unknown>,
+    }))
+  }
+
   async createPayment({
     paymentSessionId,
     payload,
   }: {
     paymentSessionId: string;
-    payload: PaymentCreateRequest;
+    payload: PostStoreMercadopagoPaymentType['paymentData'];
   }) {
     const payment = new Payment(this.client_);
     return payment.create({